+++ packages/orq-rc/src/lib/config.ts
@@ -68,7 +68,7 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 export const SDK_METADATA = {
   language: "typescript",
   openapiDocVersion: "2.0",
-  sdkVersion: "4.5.0-rc.18",
+  sdkVersion: "4.5.0-rc.19",
   genVersion: "2.845.15",
-  userAgent: "speakeasy-sdk/typescript 4.5.0-rc.18 2.845.15 2.0 @orq-ai/node",
+  userAgent: "speakeasy-sdk/typescript 4.5.0-rc.19 2.845.15 2.0 @orq-ai/node",
 } as const;